function call fixed

+++ stress_testing/dbf76cf4-8689-46a2-928f-a5474fefdbca.c
@@ -1,4 +1,4 @@
-int fucn(float x){
+int fucn(int x){
     return 1;
 }
 int main() {
